Client Hosting works with small display bugs
